+++ test/context.test.js
@@ -1,34 +1,41 @@
 const test = require('ava');
+const Koa = require('koa');
+const { request } = require('./_helper');
+const Router = require('../lib/router');
 const Route = require('../lib/route');
-const { create, request } = require('./_helper');
 
 test('exposes at ctx.router', async t => {
-  const router = create();
-  const ctx = await request(router.routes()).get('/');
+  const app = new Koa();
+  const router = new Router();
+  router.get('/', ctx => t.is(ctx.router, router));
+  app.use(router.routes());
 
-  t.is(ctx.router, router);
+  await request(app).get('/');
 });
 
 test('exposes the matched path to ctx._matchedRoute', async t => {
-  const router = create();
-  router.get('/some-path', () => {});
-  const { _matchedRoute } = await request(router.routes()).get('/some-path');
+  const app = new Koa();
+  const router = new Router();
+  router.get('/some-path', ctx => t.is(ctx._matchedRoute, '/some-path'));
+  app.use(router.routes());
 
-  t.is(_matchedRoute, '/some-path');
+  await request(app).get('/some-path');
 });
 
 test('exposes the matched route to ctx.matchedRoute', async t => {
-  const router = create();
-  router.get('/some-path', () => {});
-  const { matchedRoute } = await request(router.routes()).get('/some-path');
+  const app = new Koa();
+  const router = new Router();
+  router.get('/some-path', ctx => t.true(ctx.matchedRoute instanceof Route));
+  app.use(router.routes());
 
-  t.true(matchedRoute instanceof Route);
+  await request(app).get('/some-path');
 });
 
 test('exposes the matched route name to ctx._matchedRouteName', async t => {
-  const router = create();
-  router.get('named', '/some-path', () => {});
-  const { _matchedRouteName } = await request(router.routes()).get('/some-path');
+  const app = new Koa();
+  const router = new Router();
+  router.get('named', '/some-path', ctx => t.is(ctx._matchedRouteName, 'named'));
+  app.use(router.routes());
 
-  t.is(_matchedRouteName, 'named');
+  await request(app).get('/some-path');
 });