+++ test/invocation.test.js
@@ -1,17 +1,22 @@
 const test = require('ava');
-const { create, request } = require('./_helper');
+const Koa = require('koa');
+const { request } = require('./_helper');
+const Router = require('../lib/router');
 
 test('awaits async router handlers', async t => {
-  const router = create();
+  t.plan(2);
+  const app = new Koa();
+  const router = new Router();
+  router.get('/', async (ctx, next) => {
+    ctx.upstream = true;
+    await next();
+    t.truthy(ctx.downstream);
+  });
+  app.use(router.routes());
+  app.use(ctx => {
+    t.truthy(ctx.upstream);
+    ctx.downstream = true;
+  });
 
-  router.get('/', (ctx) => (
-    new Promise((resolve) => (
-      setTimeout(() => {
-        t.pass();
-        resolve();
-      }, 10)
-    ))
-  ));
-
-  await request(router.routes()).get('/');
+  await request(app).get('/');
 });